+++ apps/mobile/locales/en/auth.ts
@@ -1,0 +1,48 @@
+const auth = {
+  guide: {
+    welcome: 'Welcome to Imperial Kitchen',
+    slogan: 'Customize your kitchen menu easily'
+  },
+
+  createKitchen: {
+    title: 'Create Kitchen',
+    hero: 'Create Kitchen',
+    description: 'Create your kitchen now and start exploring'
+  },
+
+  joinKitchen: {
+    title: 'Join Kitchen',
+    hero: 'Join Kitchen',
+    description: 'Join a kitchen now and start exploring'
+  },
+
+  signIn: {
+    title: 'Sign In',
+    hero: 'Hi, Welcome!',
+    description: 'Get ready to dive back into the app',
+    haveNoAccount: "Don't have an account? ",
+    joinOrCreate: 'Join or Create a Kitchen'
+  },
+
+  signUp: {
+    title: 'Sign Up',
+    hero: 'Sign Up',
+    description: 'Create your account now and start exploring'
+  },
+
+  forgotPassword: {
+    title: 'Forgot Password',
+    hero: 'Reset Password',
+    description: 'Recover your account password'
+  },
+
+  createNewPassword: {
+    title: 'Create New Password',
+    hero: 'Create New Password',
+    description: 'Create your new password'
+  },
+
+  alreadyHaveAccount: 'Already have an account? '
+};
+
+export default auth;
