+++ apps/mobile/locales/zh/auth.ts
@@ -1,0 +1,47 @@
+const auth = {
+  guide: {
+    welcome: '欢迎来到御膳房',
+    slogan: '定制您的厨房菜单'
+  },
+  createKitchen: {
+    title: '创建厨房',
+    hero: '创建厨房',
+    description: '创建您的厨房并开始探索'
+  },
+
+  joinKitchen: {
+    title: '加入厨房',
+    hero: '加入厨房',
+    description: '加入厨房并开始探索'
+  },
+
+  signIn: {
+    title: '登录',
+    hero: '你好！',
+    description: '准备开始探索御膳房',
+    haveNoAccount: '没有账号？',
+    joinOrCreate: '加入或创建厨房'
+  },
+
+  signUp: {
+    title: '注册',
+    hero: '注册',
+    description: '创建您的账号并开始探索'
+  },
+
+  forgotPassword: {
+    title: '忘记密码',
+    hero: '重置密码',
+    description: '找回您的账号密码'
+  },
+
+  createNewPassword: {
+    title: '创建新密码',
+    hero: '创建新密码',
+    description: '重置您的账号密码'
+  },
+
+  alreadyHaveAccount: '已有账号？'
+};
+
+export default auth;
